More attempts to get Shared Worker working (second connect event isn't received, doesn't appear to be firing at all)
