+++ libs/plugins/src/ipc/testing/server-shared.ts
@@ -4,6 +4,7 @@ import * as exampleModule from "@doseofted/prim-example"
 
 const context = "SharedWorker"
 
+console.log("i'm shared")
 createPrimServer({
 	module: exampleModule,
 	methodHandler: createMethodHandler({ worker: self, context }),
@@ -12,3 +13,10 @@ createPrimServer({
 })
 
 export type { exampleModule }
+
+// self.addEventListener("connect", () => {
+// 	console.log("WHAT'S UP?")
+// })
+// self.onconnect = () => {
+// 	console.log("WHAT'S UP? 2")
+// }
